+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 10.3
 ---------------
 - Add "Cumulative Balance over school years" checkbox in StudentBalances.php
+- Fix program not found when query string is URL encoded in Modules.php
 
 Changes in 10.2.3
 -----------------